+++ src/MotoStore/Views/Pages/DataPagePages/UserListPage.xaml.cs
@@ -125,50 +125,51 @@ namespace MotoStore.Views.Pages.DataPagePages
             // Kiểm tra xem key Delete có được bấm trong khi đang chỉnh sửa ô hay không
             DataGridRow dgr = (DataGridRow)dg.ItemContainerGenerator.ContainerFromIndex(dg.SelectedIndex);
             if (e.Key == Key.Delete && !dgr.IsEditing)
-            {
-                // Nếu đáp ứng đủ điều kiện sẽ bắt đầu vòng lặp để xóa
-                MainDatabase mdb = new();
-                SqlCommand cmd;
-                using SqlConnection con = new(Settings.Default.ConnectionString);
-                try
+                if (MessageBox.Show("Bạn có chắc muốn xóa? Hành động này không thể hoàn tác!", "Xóa tài khoản", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
                 {
-                    con.Open();
-                    using var trans = con.BeginTransaction();
+                    // Nếu đáp ứng đủ điều kiện sẽ bắt đầu vòng lặp để xóa
+                    MainDatabase mdb = new();
+                    SqlCommand cmd;
+                    using SqlConnection con = new(Settings.Default.ConnectionString);
                     try
                     {
-                        cmd = new(" ", con, trans);
-
-                        foreach (var obj in dg.SelectedItems)
+                        con.Open();
+                        using var trans = con.BeginTransaction();
+                        try
                         {
-                            if (obj is not UserApp user)
-                                continue;
-                            // Trường hợp chưa thêm mới nên chưa có mã NV
-                            if (!mdb.UserApps.Any(nv => nv.MaNv == user.MaNv))
-                                continue;
-                            // Xóa hàng
-                            else
-                                cmd.CommandText += $"Delete From UserApp Where MaNv = '{user.MaNv}';\n";
+                            cmd = new(" ", con, trans);
+
+                            foreach (var obj in dg.SelectedItems)
+                            {
+                                if (obj is not UserApp user)
+                                    continue;
+                                // Trường hợp chưa thêm mới nên chưa có mã NV
+                                if (!mdb.UserApps.Any(nv => nv.MaNv == user.MaNv))
+                                    continue;
+                                // Xóa hàng
+                                else
+                                    cmd.CommandText += $"Delete From UserApp Where MaNv = '{user.MaNv}';\n";
+                            }
+                            cmd.ExecuteNonQuery();
+                            trans.Commit();
+                            cmd = new($"Set Dateformat dmy\nInsert into LichSuHoatDong values(newid(), '{PageChinh.getNV.MaNv}', '{DateTime.Now:dd-MM-yyyy HH:mm:ss}', N'chỉnh sửa database tài khoản')", con);
+                            cmd.ExecuteNonQuery();
                         }
-                        cmd.ExecuteNonQuery();
-                        trans.Commit();
-                        cmd = new($"Set Dateformat dmy\nInsert into LichSuHoatDong values(newid(), '{PageChinh.getNV.MaNv}', '{DateTime.Now:dd-MM-yyyy HH:mm:ss}', N'chỉnh sửa database tài khoản')", con);
-                        cmd.ExecuteNonQuery();
+                        catch (Exception ex)
+                        {
+                            trans.Rollback();
+                            MessageBox.Show(ex.Message);
+                            // Báo đã thực hiện xong event để ngăn handler mặc định cho phím này hoạt động
+                            e.Handled = true;
+                        }
                     }
                     catch (Exception ex)
                     {
-                        trans.Rollback();
                         MessageBox.Show(ex.Message);
                         // Báo đã thực hiện xong event để ngăn handler mặc định cho phím này hoạt động
                         e.Handled = true;
                     }
                 }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message);
-                    // Báo đã thực hiện xong event để ngăn handler mặc định cho phím này hoạt động
-                    e.Handled = true;
-                }
-            }
         }
 
         private void RefreshView(object sender, RoutedEventArgs e)
